Edit.ejs
Need to refactor the date code.
The Date() in js takes in ints for the date and in this order y, m, d,
h, m, s.
The html date input requires the date to be a string and this format "yyyy-mm-dd".
I split the date into separate variables and converted to strings.  Then I
created a method, normalizeDate(), that takes in a string and adds a
leading "0" if the string is of length 1.  This takes care of meeting
the "mm" and "dd" requirements when the month and day are single digits.
In the js code, if you start with a leading 0 it will assume the number
will be of a different base, oct or hex.
I then constructed a new date string to follow the "yyyy-mm-dd" format
needed for the html date.
I went with the same for the time as well just to keep the date
compnents uniform.