Added Create Route
Create is a POST under the route "/events".
I had to add some extra code to allow the date and time format from the
html date to fit into the Date of js.
The js Date() requires ints and in this order y, m, d, h, m, s.
So I split the date into separate components, converted them to ints.
Created a new Date variable using the ints and in the proper order.