Refactored Event Code
Refactored the event parse code.
The event parse code is needed due to the date and time.
The date and time are stored in the same field.
For the date to be stored correctly it needed to be parsed into an int.
The Date() js required the values to be ints and in a certain order.
  - y, m, d, h, m, s
I created a function called ParseEvent that:
  - parses the event
  - creates the proper types
  - stores the date and time in the proper order
  - returns a new object ready to store or update in the database